fix the pens ambit, add distribute
