+++ test/drawbot/style_test.py
@@ -12,12 +12,23 @@ mistral = Font.Cacheable("~/Type/fonts/fonts/_script/MistralD.otf")
 
 s = (StyledString("Hello",
     Style(mistral, 300))
-    .pen()
+    .pens()
     .f(hsl(0.9, s=1))
     .align(r))
 
-db.rect(*s.ambit())
+with db.savedState():
+    db.fill(None)
+    db.stroke(0)
+    db.strokeWidth(2)
+    db.rect(*s.ambit())
+
 s.chain(dbdraw)
+
+circle = DraftingPen().oval(r).reverse()
+(s.copy()
+    .zero_translate()
+    .distribute_on_path(circle)
+    .chain(dbdraw))
 
 db.fontSize(24)
 db.text("Mistral", s.ambit().inset(0, -50).ps, align="center")
